erratum: avoid another syncBugs() call
Similar to the change in 4e59376818c48f099ea2ade015444328412a5898, we
have to remove this call because it's not implemented in our public
codebase right now.

As described in the other syncBugs() removal change, removing this here
might cause issues for RHOS's use of this class :( Need to think about
that further.