ensure we verify SSL certificate
Prior to this change, the errata-tool lib disabled requests' SSL
verification by default. (I assume this was done for the sake of
ease-of-use, or maybe some non-prod host did not have a working cert.)

To improve security, enable SSL verification by default instead.

Add tests to ensure this behavior going forward.

Add documentation to the README for users to trust the root CA.