erratum: check if under embargo or not
The kernel team has need to query the kernel errata to check if the latest
kernel build is under embargo. Discussion with folks on the errata-tools
team revealed that this is conveyed by an erratum having the release_date
field set, with a date that is in the future. This just adds a check for
release_date, and compares it with the current date, copies the
release_date field and sets an embargoed flag in Erratum, both of which
can be consumed by scripts using this library.

CC: Don Zickus <dzickus@redhat.com>
CC: Prarit Barghava <prarit@redhat.com>
CC: Aris Rozanski <aris@redhat.com>
CC: Jeff Burke <jburke@redhat.com>
CC: Rafael Aquini <aquini@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>