Add errata call timings
When processing large numbers of errata from higher-level tools,
it's helpful to understand where the time is spent to see if multiple
calls can be avoided.

This patch adds a timings table to ErrataConnector.  Each
GET/PUT/POST is recorded, along with totals / mean / min / max.

URL APIs are deduplicated based on their name, so two calls to
different errata on the same API is recorded as a single API.

The timings only are recorded if ErrataConnector.debug is set
to True.

   { 'GET':
     { 'http://foo/***.json':
       { 'count': 10,         # number of calls to this API
         'total': 203.2,      # time spent in seconds
         'mean': 20.32,       # Average call time
         'min': 11.20,        # Fastest call
         'max': 31.3,         # Slowest call
       }
    },
    'POST': {
    ...
    },
    'PUT': {
    ...
    },
  }

To extract the information and print it, one might use PrettyTable:

    pt = PrettyTable()
    for c in ErrataConnector.timings:
        for u in ErrataConnector.timings[c]:
            pt.add_row([c, u,
                       ErrataConnector.timings[c][u]['count'],
                       ErrataConnector.timings[c][u]['total'],
                       ErrataConnector.timings[c][u]['mean'],
                       ErrataConnector.timings[c][u]['min'],
                       ErrataConnector.timings[c][u]['max']])
    print pt.get_string()

... or simply print them out.

Signed-off-by: Lon Hohberger <lhh@redhat.com>