Add product, release, and version handling
Not really an API so much as a data fetcher/aggregator
that tools can use to avoid hard-coding data normally
required to manipulate errata.

The caller is expected to cache data locally, since
this is a lot of calls and the data rarely changes.

Save/load callback functions can be specified when
instantiating, or the user can feed in the proper
data on their own.

The data is presently exposed primarly as large dicts.

Signed-off-by: Lon Hohberger <lhh@redhat.com>