setup.py: drop urllib3 dep
We're still importing this library, but apparently the newest requests
versions bundle it, and in fact setuptools breaks when we have it listed
here. (yikes). "python setup.py develop" claims it cannot find a library
named "requests". O_O

To work around the setuptools breakage, just drop it from the setuptools
packaging.