Strip severity string from synopsis when writing
Original patch by Garth Mollett <gmollett@redhat.com>. Quoting his
email:

I've noticed that when tools that use python-errata-tool make changes to
RHSA errata in ET after PS and docs approval are granted that ET always
rescinds docs and ps approval even when it shouldn't because something
like this happens:

"
synopsis changed:

@@ -1,2 +1,2 @@
-Moderate: openvswitch security update
+Moderate: Moderate: openvswitch security update
"

The root cause here is ET automatically prepends the "Severity:" part at
the start of the synopsis for RHSA's and while it's smart enough to
strip out the extra one if you add it yourself, it's not smart enough to
not count it as a docs change.

The practical upshot of this, is that if you read an erratum from ET and
then write the exact some thing back, it will cause the above behavior
("Moderate: blah" becomes "Moderate: Moderate: blah").

I've attached a simple patch that simply strips the severity from the
start of the synopsis before writing to ET if errata_type == RHSA.

Signed-off-by: Ken Dreyer <kdreyer@redhat.com>