handle PDC advisory creation API response
The REST API returns an advisory type value prefixed with "pdc_" when
creating a new advisory.

This server behavior may change in the future
(rhbz#1493773). For now, handle these gracefully and return the regular
values to callers. For example, transform "pdc_rhba" to "rhba", so
advisory.errata_type will continue to be "RHBA".