handle PDC advisory type response
The REST API returns an advisory type value prefixed with "Pdc" when
returning a PDC-based release's advisory.

This server behavior may change in the future (rhbz#1493773). For now,
handle this gracefully and return the regular value to callers. For
example, transform "PdcRHBA" to "RHBA", so advisory.errata_type will
continue to be "RHBA".