track _is_pdc, different param to /add_builds
The Errata Tool's "/add_builds" REST API endpoint requires a different
parameter, "pdc_release", instead of the older "product_version"
parameter.

I've asked the ET team to consider consolidating these two paths so
that clients will not need to know whether a release is PDC-based or
not.

In the meantime, I've added a _isPDC() private method to distinguish
whether an advisory is PDC-based or not. addBuilds() will call this
to determine the right parameter (product_version or pdc_release) to
send. The impact of this approach is that addBuilds() requires one
additional lookup per advisory object in memory. If you know whether
an advisory is pdc-based or not ahead of time, you can set
advisory._is_pdc to True or False to avoid this lookup.