connector: raise with specific HTTP 500 error message
If we get an HTTP 500 response from the Errata Tool, we probably have a
useful error message in the HTTP body.

As one example: when I try to create a new advisory for a new release
before RCM has created the new release, the ET returns an HTTP 500 with
the following JSON:

  {"error": "ERROR: Couldn't find Release without an ID"}

Instead of printing the json to STDOUT, return the data in the exception
itself.

This allows callers to discover the specific error message. Callers can
print human-readable explanations about what went wrong and what steps
to take next (for example, "file an RCM ticket to create this new
release").