refactor metadataCdnRepos to _cdn_repos
Move the bulk of the metadataCdnRepos() logic to a private _cdn_repos()
method that takes an API endpoint argument.

The goal of this change is to make it easier to add support for
text-only advisories' repo metadata. Text-only advisories use almost
exactly the same logic, but a slightly different REST API endpoint URL.