tests: update fixture values
Refresh almost all the test fixtures to use more modern values.

Prior to this change, we were relying on some older release and product
version configurations, and these are no longer valid now that PDC has
gone away.

With this change, we do not refer to PDC at all in the test fixtures.