connector: better debugging for HTTP 500 GET errors
Prior to this change, if we GET a URL and received a HTTP 500 response,
we would raise a simple LookupError with no further information about
what went wrong.

Update the error handling logic in _get() so that we will raise
ErrataException instead. This will give us a lot more information about
the URL we were trying to access and the response back from the server.