Update strptime format to include millseconds
This is a minimum viable patch to support the timestamp format
change (possibly) coming in Errata Tool 3.20.9.

I haven't tested it. Note that it is not backward compatible, so it
would require Errata Tool 3.20.9 or higher.

Note also that it's been suggested that we might change the
timestamp format back to the old format in ET 3.20.9. If that
happens then this patch must not be merged.

Github: #187
JIRA: ERRATA-9192