tests: fix release.blocker_flags test
Prior to this change, pytest was not invoking the blocker_flags() method
because it did not start with "test_". Rename the method so that pytest
invokes it.

Our "expected" values were ordered incorrectly when compared to our test
fixture data. Re-order the expected flags so that "pm_ack" is the final
one.

With these changes, pytest properly exercises .blocker_flags attribute.