erratum: get paginated data in externalTests()
Prior to this change, we would only return the first page of
external_test results. For advisories with many packages, this means
that we could miss RPMDiff failures further down the list.

Introduce a new "get_paginated_data()" method on the Connector class.
This method will query an API endpoint and return all the data from all
pages.