tests: load test fixtures according to query string parameters
Prior to this change, the test suite's MockResponse class only loaded
test fixtures according to the URLs, without regard for the request
params kwarg.

Pass the kwarg through CallRecorder and MockRequests, and use this to
determine the final filename of the HTTP request fixture.

With this change, we can write unit tests that exercise the new "params"
kwarg to _get().