release: get release with requests' params
Don't manually assemble the GET request filter URL. Use requests' params
kwarg so that we quote special characters - particularly the "+"
character.

This fixes a bug where we would load the wrong URLs for releases with
"+" in the names. When we loaded URLs with unquoted "+" characters in
the query strings, the ET web server interprets those as spaces, so it
would return no result, and we would raise NoReleaseFoundError.

With this change, we properly encode all query string parameters for
releases, in particular "+".