erratum: fix updating owner and manager emails
Prior to this commit, users could not update owner and manager email
addresses for existing advisories. We sent the parameters when creating
a new advisory, but not when editing an existing advisory.

Move the pdata assignments out of the "if self._new" conditional so that
we send these parameters during advisory creations and updates.